pyproject.toml
[tool.poetry]
name = "automona"
version = "0.1.0"
description = "Automona is a powerful tool designed to automate tasks related to performing daily activities."
authors = ["Gabriel Mango <gabrielhmango@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.16.0"
webdriver-manager = "^4.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pip-audit = "^2.6.2"
isort = "^5.13.2"
mkdocs = "^1.5.3"
blue = "^0.9.1"
pylint = "^3.0.3"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
